Examples fixed
